Fixed bug in completions
Fixed bug: For a given word completion, the number of hits should be
the number of matching documents with the completion. Instead, it is
equal to the number of occurrences of the completion on the matching
documents. Because of this, the number of hits per completion is higher
than expected.